requirements.txt
Flask==2.3.3
python-docx==0.8.11
pytenable==1.4.22
defusedxml==0.7.1
Werkzeug==2.3.7
MarkupSafe==2.1.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
lxml==4.9.3 